Another batch of changes
Added a system to combine nether quartz gems, rather than dusts, to make it harder to make a free barium loop.

Removed all of the ores from the dissolver. These should be processed to dusts before being dissolved, and because I don't want to deal with the calculation complexities that can result from allowing them.