Another large batch of work.
Centraize a few scripts to make things cleaner. The recipes for the compactor, dissolver and combiner for all dusts available by ChemLib are handled automatically.

Introduce an alias system to reduce the number of recipe files needed. This is used where there are materials with the same chemical composition and don't have any special bonus outputs.